requirements.txt
python-telegram-bot==13.7
Pillow==9.5.0
ffmpeg-python==0.2.0
PyPDF2==3.0.1
reportlab==4.0.4
